refactor: move shared config functionality back to APNS.Shared
Necessary to clean up compile warnings.
